+++ Cpp/constructors/p2.cpp
