--- Cpp/p5.cpp
@@ -1,1 +1,0 @@
-
