+++ classTask/Random/1.cpp
