+++ classTask/Random/3.cpp
