+++ classTask/Random/friendclass.cpp
