+++ classTask/Random/p3.cpp
