+++ classTask/Random/sum.cpp
