+++ classTask/Random/valueTansfer.cpp
