--- classTask/class.cpp
@@ -1,31 +1,0 @@
-#include<iostream>
-using namespace std;
-
-class X{
-	private:
-	int id;
-	char studentName[25];
-	char studentAddress[25];
-	char studentFaculty[4];
-	public:
-	void name()
-	{
-	
-		studentName[25]=a[25];
-	}
-	void address()
-	{
-		studentAddress[25]=b[25];
-	}
-	void faculty()
-	{
-		studentFaculty[4]=c[4];
-	}
-	
-};
-int main(){
-	char a[25];
-	char b[25];
-	char c[4];
-	
-}
