--- classTask/classTask_1.txt
@@ -1,3 +1,0 @@
-WAP to add two object of a class using friend function the class consist of hour and minute as its data member,input() ,sum() and output() 
-WAP to add two integers and return the sum using inline function. (2017 Q2.b)
-WAP to add two complex number of two different class.
