--- trash/Add-number-using-function.cpp
@@ -1,19 +1,0 @@
-// C++ program to add two number
-#include <iostream>
-using namespace std;
-
-int addTwoNumber(int A, int B)
-{
-
-	return A + B;
-}
-
-int main()
-{
-
-	int A = 4, B = 11;
-
-	cout << "sum = " << addTwoNumber(A, B);
-	return 0;
-}
-
