+++ trash/new_1.cpp
